Popups relationship
A pop-up will show up if whether the relationship was created successfully or not.

The form of the relationship association contains a picklist with all the nivs of the vehicles that don't have a relationship with a cellphone yet.
By doing this we make sure that a cellphone can only be related with one vehicle